requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
torch==2.1.0
numpy==1.26.2
matplotlib==3.8.2
